% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPSclass.R
\docType{methods}
\name{bhai}
\alias{bhai}
\alias{bhai,PPS-method}
\alias{bhai,PPS,ANY-method}
\title{Main function of the package to estimation of the burden of healthcare-associated infections}
\usage{
bhai(pps, nsim = 1000, pop.sampling = TRUE,
  sample_distr = "rbetamix", estimate_loi_fun = bootstrap_mean_gren,
  stratified_sampling = FALSE, summarize_strata = TRUE,
  use_prior = TRUE)

\S4method{bhai}{PPS}(pps, nsim = 1000, pop.sampling = TRUE,
  sample_distr = "rbetamix", estimate_loi_fun = bootstrap_mean_gren,
  stratified_sampling = FALSE, summarize_strata = TRUE,
  use_prior = TRUE)
}
\arguments{
\item{pps}{The PPS object containing the data.}

\item{nsim}{Number of Monte Carlo simulations, default: 1000.}

\item{pop.sampling}{Specifying whether parameters of the disease outcome trees should be sampled on population level, default: TRUE.}

\item{sample_distr}{Distribution used for prevalence sampling, default: "rbetamix".}

\item{estimate_loi_fun}{Function used for estimation of the length of infection, default: bootstrap_mean_gren (recommended!).}

\item{stratified_sampling}{Specifying whether stratified sampling should be done.}

\item{summarize_strata}{Specifying whether stratum-specific summary statistics should be computed.}

\item{use_prior}{Specifying whether Prior distributions should be used for computations.}
}
\value{
A PPS class object.
}
\description{
Estimation of the burden of healthcare-associated infections
}
\examples{

data(german_pps_2011_repr)
german_pps_repr = PPS(num_hai_patients = num_hai_patients,
    num_hai_patients_by_stratum = num_hai_patients_by_stratum,
    num_hai_patients_by_stratum_prior = num_hai_patients_by_stratum_prior,
    num_survey_patients = num_survey_patients,
    length_of_stay = length_of_stay,
    loi_pps = loi_pps,
    mccabe_scores_distr = mccabe_scores_distr,
    mccabe_life_exp = mccabe_life_exp,
    hospital_discharges = hospital_discharges,
    population = population,
    country="Germany (representative sample)")
german_pps_repr

set.seed(3)
# The following example is run only for illustratory reasons
# Note that you should never run the function with only 10 Monte-Carlo simulations in practice!
bhai(german_pps_repr, nsim=10)

}
\seealso{
\code{\linkS4class{PPS}}
}
