% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{eu_pps}
\alias{eu_pps}
\title{Aggregated data of the ECDC PPS 2010-2011.}
\format{A PPS object.}
\usage{
data(eu_pps_2011)
}
\description{
Aggregated data of the ECDC PPS 2010-2011.
}
\keyword{datasets}
