% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{german_pps_conv}
\alias{german_pps_conv}
\title{Aggregated data of the german PPS 2010-2011 (convenience sample).}
\format{A PPS object.}
\usage{
data(german_pps_2011_conv)
}
\description{
Aggregated data of the german PPS 2010-2011 (convenience sample).
}
\keyword{datasets}
