% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{hospital_discharges}
\alias{hospital_discharges}
\title{Hospital discharges in Germany (2011)}
\format{A PPS object.}
\usage{
data(german_pps_2011_repr)
}
\description{
Hospital discharges in Germany (2011)
}
\keyword{datasets}
