% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{length_of_stay}
\alias{length_of_stay}
\title{Average length of stay of survey patients in german PPS 2011 (representative sample)}
\format{A PPS object.}
\usage{
data(german_pps_2011_repr)
}
\description{
Average length of stay of survey patients in german PPS 2011 (representative sample)
}
\keyword{datasets}
