% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{loi_pps}
\alias{loi_pps}
\title{A list containing length of infections from all patients in the german PPS 2011 representative sample.}
\format{A PPS object.}
\usage{
data(german_pps_2011_repr)
}
\description{
A list containing length of infections from all patients in the german PPS 2011 representative sample.
}
\keyword{datasets}
