% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{mccabe_life_exp}
\alias{mccabe_life_exp}
\title{Named list containing remaining life expectancies for each McCabe score (NONFATAL, ULTFATAL, RAPFATAL).}
\format{A PPS object.}
\usage{
data(german_pps_2011_repr)
}
\description{
Named list containing remaining life expectancies for each McCabe score (NONFATAL, ULTFATAL, RAPFATAL).
}
\keyword{datasets}
