% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{mccabe_scores_distr}
\alias{mccabe_scores_distr}
\title{The observed McCabe scores (counts) for each infection, age and gender stratum from the ECDC PPS 2011-2012.}
\format{A PPS object.}
\usage{
data(german_pps_2011_repr)
}
\description{
The observed McCabe scores (counts) for each infection, age and gender stratum from the ECDC PPS 2011-2012.
}
\keyword{datasets}
