% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{num_hai_patients_by_stratum_prior}
\alias{num_hai_patients_by_stratum_prior}
\title{Stratified number of cases for each infection in the german PPS 2011 (convenience sample). This distribution is used as a Prior for the representative sample.}
\format{A PPS object.}
\usage{
data(german_pps_2011_repr)
}
\description{
Stratified number of cases for each infection in the german PPS 2011 (convenience sample). This distribution is used as a Prior for the representative sample.
}
\keyword{datasets}
