% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{num_survey_patients}
\alias{num_survey_patients}
\title{Number of survey patients in the german PPS 2011 (representative sample).}
\format{A PPS object.}
\usage{
data(german_pps_2011_repr)
}
\description{
Number of survey patients in the german PPS 2011 (representative sample).
}
\keyword{datasets}
