% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{population}
\alias{population}
\title{Population size of Germany in 2011.}
\format{A PPS object.}
\usage{
data(german_pps_2011_repr)
}
\description{
Population size of Germany in 2011.
}
\keyword{datasets}
