% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.pps.R
\name{sample.pps}
\alias{sample.pps}
\title{Simulate PPS data}
\usage{
sample.pps(pps_data, num_survey_patients)
}
\arguments{
\item{pps_data}{The PPS object containing the data. Parameters for simulations are extracted from this data.}

\item{num_survey_patients}{Numeric vector indicating sample sizes for simulations.}
}
\value{
A simulated PPS object.
}
\description{
Simulate PPS data
}
\examples{

# Specify the number of survey patients
sim_survey_patients = 10000
# Subsample data sets from european PPS
sim_pps = sample.pps(eu_pps, num_survey_patients = sim_survey_patients)

}
\seealso{
\code{\linkS4class{PPS}}
}
