% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{sim_pps}
\alias{sim_pps}
\title{Simulated/subsampled data sets from european PPS}
\format{A PPS object.}
\usage{
data(simulations)
}
\description{
Simulated/subsampled data sets from european PPS
}
\keyword{datasets}
