% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAI-package.R
\docType{data}
\name{sim_pps_bhai_prior}
\alias{sim_pps_bhai_prior}
\title{BHAI with prior was applied to simulated/subsampled data sets from european PPS}
\format{A PPS object.}
\usage{
data(simulations)
}
\description{
BHAI with prior was applied to simulated/subsampled data sets from european PPS
}
\keyword{datasets}
